cmd/compile: shrink binary
DO NOT REVIEW

has no effect :(

todo: measure, etc.

+++ src/cmd/compile/internal/ssa/check.go
@@ -126,7 +126,7 @@ func checkFunc(f *Func) {
 			// Check to make sure argument count makes sense (argLen of -1 indicates
 			// variable length args)
 			nArgs := opcodeTable[v.Op].argLen
-			if nArgs != -1 && int32(len(v.Args)) != nArgs {
+			if nArgs != -1 && int8(len(v.Args)) != nArgs {
 				f.Fatalf("value %s has %d args, expected %d", v.LongString(),
 					len(v.Args), nArgs)
 			}
